refactor: 상태를 사용하지 않는 url 관련 함수들 분리
